requirements.txt
streamlit
requests
geopy
sqlite3-binary
pandas
python-dotenv